backend/scraper-api/requirements.txt
Flask==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
openai==0.27.8
python-dotenv==1.0.0
flask-cors==3.0.10